+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
+### MASTER
+### Fixed
+- Removed the element option's erroneous "all" value from `backup:get`, changed its default to "files". (#1705)
+
 ## 1.3.0 - 2017-04-20
 ### Added
 - `env:commit` now has a `--force` option to force a commit even if no changes are found. (#1115)